pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Orion"
version = "0.1.0"
description = "A local multi-user Personal Knowledge RAG using Ollama"
authors = [
    { name="Ralf Wadee L. Hernandez", email="ralf.hernandez090102@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core RAG & LLM
    "ollama",
    "chardet",
    "rank_bm25",
    "langchain",
    "langchain-community", 
    "langchain-ollama",
    "langchain-text-splitters",
    "sentence_transformers",
    "faiss-cpu",
    
    # Document processing
    "pandas",
    "python-docx",
    "python-pptx",
    "openpyxl",
    "pypdf",
    "PyMuPDF",
    "docx2txt",
    "pillow",
    
    # FastAPI Backend
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    
    # System integration
    "pystray",
    "psutil",
    
    # Utilities
    "tqdm",
    "typer",
    "rich",
    "requests"
]

[tool.setuptools]
packages = ["core", "backend", "system_tray"]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest",
    "pre-commit"
]
media = [
    "easyocr",
    "pytesseract", 
    "camelot-py[cv]",
    "tabula-py"
]


[tool.black]
line-length = 88
target-version = ['py313']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

[tool.ruff]
line-length = 128

[tool.ruff.lint]
select = ["E", "F"]

[tool.pytest.ini_options]
testpaths = ["tests"]
